fix(queries): check for invalid ObjectId

+++ packages/strapi-connector-mongoose/lib/queries.js
@@ -14,6 +14,7 @@ const populateQueries = require('./utils/populate-queries');
 const { PUBLISHED_AT_ATTRIBUTE, DP_PUB_STATES } = contentTypesUtils.constants;
 const { findComponentByGlobalId } = require('./utils/helpers');
 const { handleDatabaseError } = require('./utils/errors');
+const utils = require('./utils')();
 
 const hasPK = (obj, model) => _.has(obj, model.primaryKey) || _.has(obj, 'id');
 const getPK = (obj, model) => (_.has(obj, model.primaryKey) ? obj[model.primaryKey] : obj.id);
@@ -65,6 +66,14 @@ module.exports = ({ model, strapi }) => {
     } catch (error) {
       return handleDatabaseError(error);
     }
+  };
+
+  const isValidPrimaryKey = params => {
+    if (_.has(params, model.primaryKey)) {
+      return utils.isMongoId(params[model.primaryKey]);
+    }
+
+    return true;
   };
 
   async function createComponents(entry, values, { isDraft, session = null } = {}) {
@@ -435,6 +444,10 @@ module.exports = ({ model, strapi }) => {
   }
 
   async function findOne(params, populate, { session = null } = {}) {
+    if (!isValidPrimaryKey(params)) {
+      return null;
+    }
+
     const entries = await find({ ...params, _limit: 1 }, populate, { session });
     return entries[0] || null;
   }
@@ -477,6 +490,10 @@ module.exports = ({ model, strapi }) => {
   }
 
   async function update(params, values, { session = null } = {}) {
+    if (!isValidPrimaryKey(params)) {
+      return null;
+    }
+
     const entry = await model.findOne(params).session(session);
 
     if (!entry) {
@@ -500,11 +517,27 @@ module.exports = ({ model, strapi }) => {
 
   async function deleteMany(params, { session = null } = {}) {
     if (params[model.primaryKey]) {
+      if (!utils.isMongoId(params[model.primaryKey])) {
+        return null;
+      }
+
       const entries = await find({ ...params, _limit: 1 }, null, { session });
       if (entries.length > 0) {
         return deleteOne(entries[0][model.primaryKey], { session });
       }
       return null;
+    }
+
+    const { _where = [] } = params;
+
+    for (let i = 0; i < _where.length; i++) {
+      if (_.has(_where[i], `${model.primaryKey}_in`)) {
+        params._where[i][`${model.primaryKey}_in`] = params._where[i][
+          `${model.primaryKey}_in`
+        ].filter(id => utils.isMongoId(id));
+
+        break;
+      }
     }
 
     const entries = await find(params, null, { session });
